EngineeringTable added

+++ ConfigTool.UI/ViewModels/TableViewModels/Implementations/LayerSideTableViewModel.cs
@@ -30,7 +30,7 @@ namespace ConfigTool.UI.ViewModels
                 }
 
                 //Check all columns with 'number' datatype
-                if (Int32.TryParse(TableFilter, out int result))
+                if (Int16.TryParse(TableFilter, out short result))
                 {
                     return tableItem.Table.Id.Equals(result);
                 }
